добавление папки с интерфейсами, удаление лишнего
