--- src/app/interfaces.ts
@@ -1,36 +1,0 @@
-export interface User {
-  id: number;
-  name: string;
-  username?: string;
-  email: string;
-  address?: {
-    street?: string;
-    suite?: string;
-    city?: string;
-    zipcode?: string;
-    geo?: {
-      lat?: string;
-      lng?: string;
-    };
-  };
-  phone?: string;
-  website: string;
-  company: {
-    name: string;
-    catchPhrase?: string;
-    bs?: string;
-  };
-}
-export interface CreateUserI {
-  id: number;
-  name: string;
-  email: string;
-  website: string;
-  companyName: string;
-}
-export interface Todo {
-  userId: number;
-  id: number;
-  title: string;
-  completed: boolean;
-}
